Break apart types, data, and UI into distinct files.
