+++ src/views/Prototype508/index.tsx
@@ -13,186 +13,10 @@ import { DateTime } from 'luxon';
 import BreadcrumbNav from 'components/BreadcrumbNav';
 import Modal from 'components/Modal';
 
+import projects from './data';
+import { Activity, DocumentType, ProjectStatus } from './types';
+
 import './index.scss';
-
-enum DocumentType {
-  TestPlan,
-  TestingVPAT,
-  TestResults,
-  AwardedVPAT,
-  RemediationPlan
-}
-
-type Document = {
-  type: DocumentType;
-  mimetype: string;
-  createdAt: DateTime;
-};
-
-enum ProjectStatus {
-  ConsultRequested,
-  TestingRequested,
-  TestingInProgress,
-  TestingCompleted,
-  InRemediation
-}
-
-type BusinessOwner = {
-  name: string;
-};
-
-enum ActivityType {
-  NoteAdded,
-  StatusChanged,
-  DocumentAdded,
-  DocumentRemoved,
-  ProjectCreated
-}
-
-export type Activity = {
-  authorName: string;
-  content: string;
-  createdAt: DateTime;
-  type: ActivityType;
-};
-
-type Project = {
-  id: number;
-  name: string;
-  status: ProjectStatus;
-  businessOwner: BusinessOwner;
-  submissionDate: DateTime;
-  lastUpdatedAt: DateTime;
-  lifecycleID: string;
-  description?: string;
-  documents: Document[];
-  activities: Activity[];
-};
-
-// TODO needs to be at least 15
-const projects: Project[] = [
-  {
-    id: 1,
-    name: 'TACO',
-    status: ProjectStatus.ConsultRequested,
-    submissionDate: DateTime.fromISO('2020-03-21'),
-    lastUpdatedAt: DateTime.fromISO('2020-05-17'),
-    businessOwner: {
-      name: 'Shane Clark'
-    },
-    lifecycleID: 'X200943',
-    documents: [
-      {
-        type: DocumentType.AwardedVPAT,
-        mimetype: 'application/pdf',
-        createdAt: DateTime.fromISO('2020-02-09')
-      },
-      {
-        type: DocumentType.TestPlan,
-        mimetype: 'application/pdf',
-        createdAt: DateTime.fromISO('2020-02-09')
-      },
-      {
-        type: DocumentType.AwardedVPAT,
-        mimetype: 'application/pdf',
-        createdAt: DateTime.fromISO('2020-02-09')
-      }
-    ],
-    activities: [
-      {
-        content: 'Status changed to Testing in progress',
-        createdAt: DateTime.fromISO('2020-02-09'),
-        type: ActivityType.StatusChanged,
-        authorName: 'Aaron Allen'
-      },
-      {
-        content: 'VPAT uploaded',
-        createdAt: DateTime.fromISO('2020-02-09'),
-        type: ActivityType.DocumentAdded,
-        authorName: 'Aaron Allen'
-      },
-      {
-        content: 'Test plan uploaded',
-        createdAt: DateTime.fromISO('2020-02-09'),
-        type: ActivityType.DocumentAdded,
-        authorName: 'Aaron Allen'
-      },
-      {
-        content:
-          'We are waiting on the test plan and VPAT from business owner.',
-        createdAt: DateTime.fromISO('2020-02-09'),
-        type: ActivityType.NoteAdded,
-        authorName: 'Aaron Allen'
-      },
-      {
-        content: 'Awarded VPAT uploaded',
-        createdAt: DateTime.fromISO('2020-02-09'),
-        type: ActivityType.DocumentAdded,
-        authorName: 'Aaron Allen'
-      },
-      {
-        content: 'TACO project created',
-        createdAt: DateTime.fromISO('2020-02-09'),
-        type: ActivityType.ProjectCreated,
-        authorName: 'Aaron Allen'
-      }
-    ],
-    description:
-      'TACO is a new tool for customers to access consolidated Active health information and facilitate the new Medicare process. The purpose is to provide a more integrated and unified customer service experience.'
-  },
-  {
-    id: 2,
-    name: 'Impact Analysis Network',
-    status: ProjectStatus.TestingRequested,
-    submissionDate: DateTime.fromISO('2020-03-21'),
-    lastUpdatedAt: DateTime.fromISO('2020-05-17'),
-    businessOwner: {
-      name: 'Shane Clark'
-    },
-    activities: [],
-    documents: [],
-    lifecycleID: 'X200943'
-  },
-  {
-    id: 3,
-    name: 'Migration Pipeline',
-    status: ProjectStatus.TestingInProgress,
-    submissionDate: DateTime.fromISO('2020-03-21'),
-    lastUpdatedAt: DateTime.fromISO('2020-05-17'),
-    businessOwner: {
-      name: 'Connie Leonard'
-    },
-    activities: [],
-    documents: [],
-    lifecycleID: 'X200943'
-  },
-  {
-    id: 4,
-    name: '(USDS) Dashboard for USDS',
-    status: ProjectStatus.TestingCompleted,
-    submissionDate: DateTime.fromISO('2020-03-21'),
-    lastUpdatedAt: DateTime.fromISO('2020-05-17'),
-    businessOwner: {
-      name: 'Ada Sanchez'
-    },
-    activities: [],
-    documents: [],
-    lifecycleID: 'X200943'
-  },
-  {
-    id: 5,
-    name: 'OSORA FOIA Portal Project',
-    status: ProjectStatus.InRemediation,
-    submissionDate: DateTime.fromISO('2020-03-21'),
-    lastUpdatedAt: DateTime.fromISO('2020-05-17'),
-    businessOwner: {
-      name: 'Amanda Johnson'
-    },
-    activities: [],
-    documents: [],
-    lifecycleID: 'X200943'
-  }
-];
 
 const Header = () => {
   return (
